checkstyle tweaks
