+++ service/src/main/java/io/nuvalence/platform/notification/service/utils/XmlUtils.java
@@ -1,0 +1,45 @@
+package io.nuvalence.platform.notification.service.utils;
+
+import org.jdom2.Document;
+import org.jdom2.JDOMException;
+import org.jdom2.input.SAXBuilder;
+import org.jdom2.output.Format;
+import org.jdom2.output.XMLOutputter;
+
+import java.io.IOException;
+import java.io.StringReader;
+
+import javax.xml.XMLConstants;
+
+/**
+ * Utility class for XML operations.
+ */
+public class XmlUtils {
+
+    private XmlUtils() {}
+
+    /**
+    * <p>
+    * Verifies the string represents a valid XML file.
+    * And returns it minified for transfer convenience.
+    * </p>
+    * 
+    * @param xmlString the XML string to be parsed
+    * @return valid XML string in pretty format
+    * @throws JDOMException when errors occur in parsing
+    * @throws IOException   when an I/O error prevents a document from being fully
+    *                       parsed
+    */
+    public static String xmlValidateAndMinify(String xmlString) throws JDOMException, IOException {
+
+        SAXBuilder saxBuilder = new SAXBuilder();
+        saxBuilder.setProperty(XMLConstants.ACCESS_EXTERNAL_DTD, "");
+        saxBuilder.setProperty(XMLConstants.ACCESS_EXTERNAL_SCHEMA, "");
+
+        var reader = new StringReader(xmlString);
+        Document xmlDoc = saxBuilder.build(reader);
+        reader.close();
+
+        return new XMLOutputter(Format.getCompactFormat()).outputString(xmlDoc);
+    }
+}
